Refactored CMake configuration by adding subdirectories for GameplaySystem and InputDevices; updated .gitignore to exclude .vscode; modified tasks.json to include new CMake build tasks; improved includes in various source files for better organization and clarity.

+++ sources/Misc/DataStream.h
@@ -25,10 +25,11 @@
 #pragma once
 
 #include "Configs.h"
+#include "Core/IntrusivePtr.h"
 #include "Core/String.h"
-#include "Misc/JsonAdapter.h"
+#include "JustReflectMe/Adapter.h"
+#include "nlohmann/json.hpp"
 
-#include <Core/IntrusivePtr.h>
 #include <functional>
 
 namespace Core
@@ -37,7 +38,7 @@ namespace Core
 
     struct IDataUpdateBridge
     {
-        enum Rule : uint32_t
+        enum Rule : uint8_t
         {
             None = 0,
             NoRead = 1 << 1,
